move createUserToken to utils
