+++ routers/user.js
@@ -1,5 +1,6 @@
 const { Router } = require("express")
 const { User } = require("../models/user")
+const { createUserToken } = require("../utils/auth")
 
 const router = Router()
